test: abstract utils
