+++ test/color-scheme-switch-element.test.js
@@ -1,78 +1,9 @@
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
 import { ColorSchemeSwitchElement } from '../src/color-scheme-switch-element.js';
+import { createMatchMediaController, setPersistedValue, mountElement, createElement } from './utils.js';
 
 const TAG_NAME = 'test-color-scheme-switch';
 const STORAGE_KEY = 'color-scheme';
-const DARK_MEDIA_QUERY = '(prefers-color-scheme: dark)';
-
-function createMatchMediaController(initialMatches = false) {
-  let matches = initialMatches;
-  const listeners = new Set();
-
-  const mediaQueryList = {
-    get matches() {
-      return matches;
-    },
-
-    media: DARK_MEDIA_QUERY,
-
-    addEventListener(type, listener) {
-      if (type === 'change' && listener) {
-        listeners.add(listener);
-      }
-    },
-
-    removeEventListener(type, listener) {
-      if (type === 'change') {
-        listeners.delete(listener);
-      }
-    },
-
-    addListener() {},
-    removeListener() {},
-  };
-
-  return {
-    mediaQueryList,
-
-    setMatches(nextMatches) {
-      matches = nextMatches;
-
-      const event = {
-        matches: nextMatches,
-        media: DARK_MEDIA_QUERY,
-      };
-
-      for (const listener of [...listeners]) {
-        listener(event);
-      }
-    },
-  };
-}
-
-function createElement(attributes = {}) {
-  const element = document.createElement(TAG_NAME);
-
-  for (const [name, value] of Object.entries(attributes)) {
-    if (value !== undefined) {
-      element.setAttribute(name, value);
-    }
-  }
-
-  return element;
-}
-
-function mountElement(attributes = {}) {
-  const element = createElement(attributes);
-
-  document.body.append(element);
-
-  return element;
-}
-
-function setPersistedValue(value) {
-  localStorage.setItem(STORAGE_KEY, value);
-}
 
 let matchMediaController;
